api/requirements.txt
# ============================
# Web / API
# ============================
fastapi==0.110.3
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.6.4
python-dotenv==1.0.1
aiofiles==23.2.1

# ============================
# DB (Async MySQL)
# ============================
sqlalchemy==2.0.28
aiomysql==0.2.0
pymysql==1.1.0
greenlet==3.0.3

# ============================
# OpenAI
# ============================
openai==1.30.1

# ============================
# PDF 처리
# ============================
pypdf==4.2.0

# ============================
# LangChain + FAISS (RAG)
# ============================
langchain==0.1.16
langchain-community==0.0.36
langchain-openai==0.1.7
langchain-text-splitters==0.0.1
faiss-cpu==1.8.0
httpx==0.27.2
